# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: session-broker
  version: 0.1.0
  summary: Brokers browser sessions against an upstream OIDC provider and takes custody of the upstream tokens.
  description: |
    Three properties define the service, in priority order:

    1. **Non-invalidating rotation.** Issuing a new session cookie does not
       invalidate the previous one; generations overlap for a grace window.
       This removes the refresh race class outright rather than coordinating
       around it.
    2. **Local-only fast refresh.** Session lifetime is decoupled from upstream
       token lifetime. `POST /session/refresh` performs no network I/O (INV-8).
    3. **Server-side guarantees, dumb client.** The browser SDK needs no
       cross-tab coordination beyond one Web Lock.

    ## Cookies

    - `__Host-broker_session` (`broker_session` over loopback) — `HttpOnly`,
      `SameSite=Lax`. The credential. Never readable by script.
    - `broker_meta` — deliberately readable by script (INV-9), and deliberately
      **unsigned**. It is a client hint. The server never reads it back, and
      signing it would imply it could be trusted as input.

    ## CSRF

    Same-origin path-mounted deployment is the default (INV-10, ADR-0012), so
    there is no CORS surface. The CSRF boundary is `Sec-Fetch-Site` / `Origin`
    on unsafe methods (ADR-0004), which the browser sets and script cannot forge.
  license:
    name: MIT OR Apache-2.0
servers:
  - url: https://app.example.com
...
paths:
...
  /internal/token:
    post:
      tags: [internal]
      summary: Exchange a session token for the custodied upstream access token.
      description: |
        **Dual authentication, and neither half is sufficient** (INV-11, ADR-0007):

        - a broker API key proves the caller is a trusted backend — on its own it
          cannot name a user;
        - a live session token proves *which* user — on its own it yields nothing,
          which is what stops a cookie stolen by XSS from being exchanged, since
          the browser does not hold the key.

        **No inline refresh.** If the custodied access token has expired this
        answers `custody_unavailable` rather than calling the IdP (INV-8): a
        backend seeing it should retry, not treat it as an auth failure.

        Two kinds of key work here. Keys **issued** through `/admin/api-keys` are
        the normal case and can be revoked without a restart. The single
        **bootstrap** key from config exists so the first issued key can be
        created, and cannot be revoked without a deploy.
      security:
        - brokerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_token]
              properties:
                session_token:
                  type: string
                  description: The raw session cookie value, forwarded by the backend.
      responses:
        "200":
          description: The current upstream access token.
          content:
            application/json:
              schema:
                type: object
                required: [access_token, sub, expires_at]
                properties:
                  access_token: { type: string }
                  sub:
                    type: string
                    description: The IdP subject. Opaque — map it to your own user model.
                  expires_at: { type: integer, description: Unix seconds. }
                  scope: { type: string, nullable: true }
        "401":
          description: |
            `missing_api_key` / `bad_api_key` — the CALLER's credential is wrong;
            an operator problem, not the user's.
            `session_not_active` — the session does not authenticate resource access.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LaneError" }
        "501":
          description: "`lane_disabled` — no broker API keys are configured."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LaneError" }
        "503":
          description: |
            `upstream_revoked` — the grant behind this session is dead.
            `custody_unavailable` — the access token has expired and has not yet
            been renewed by the keepalive worker. **Retry; do not log the user out.**
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LaneError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-broker_session
      description: "`HttpOnly`. Sent automatically on same-origin requests; script can never read it."
    brokerApiKey:
      type: http
      scheme: bearer
      description: |
        A backend API key — issued via `/admin/api-keys`, or the bootstrap key
        from config. Required IN ADDITION to a session token; neither alone is
        sufficient.
    adminApiKey:
      type: http
      scheme: bearer
      description: |
        The admin key. Deliberately a DIFFERENT value from the backend key: one
        that can mint keys is strictly more powerful than one that exchanges a
        session for a token, and sharing them would let every backend service
        issue credentials for itself.
...
  schemas:
...
    LaneError:
      type: object
      description: The internal and admin lanes' error shape.
      required: [error, detail]
      properties:
        error:
          type: string
          description: A stable machine-readable code. Clients should branch on this, not on the status alone.
        detail: { type: string }
security:
  - sessionCookie: []